feat: simple color correction, refactor binary

+++ src/lib.rs
@@ -14,7 +14,7 @@ pub type Image = ImageBuffer<Rgb<u8>, Vec<u8>>;
 
 /// Generic method to generate a lut using anything that implements [`InterpolatedRemapper`]
 pub fn generate_lut<'a, A: InterpolatedRemapper<'a>>(
-    level: u32,
+    level: u8,
     palette: &'a [[u8; 3]],
     params: A::Params,
 ) -> Image {
